# Optimize IndexOfFirstMismatch