@jnyrup 2020-04-18
I'm no encoding expert (at all), so I tried enumerating all 16-bit chars to verify if there was any difference.
Spoiler: there wasn't.

```c#
var chars = Enumerable.Range(char.MinValue, char.MaxValue + 1)
            .Select(i => (char)i)
            .Where(c => !char.IsControl(c))
            .ToArray();


foreach (var c1 in chars)
{
    foreach (var c2 in chars)
    {
        var comp1 = char.ToUpperInvariant(c1) == char.ToUpperInvariant(c2);
        var comp2 = string.Equals(c1.ToString(), c2.ToString(), StringComparison.OrdinalIgnoreCase);

        if (comp1 ^ comp2)
        {
            ; // <--  My breakpoint here was never hit.
        }
    }
}
```